requirements.txt
django
django-translatable
pytz
python-dateutil
Pillow
pyserial
# pyudev
uwsgi
paho-mqtt